requirements.txt
# Core dependencies
numpy>=1.22.0
pyyaml>=6.0
colorama>=0.4.6

# Optional logging enhancement
loguru>=0.7.2

# Progress bars and testing
tqdm>=4.66.0
pytest>=7.4.0

openai>=0.27.0
simpleaudio>=1.0.4
requests>=2.31.0